[elf][fpu_memcpy] Replace ptrdiff_t with long

+++ include/tilck/kernel/elf_utils.h
@@ -15,7 +15,7 @@ struct elf_symbol_info {
 
 void call_kernel_global_ctors(void);
 ulong find_addr_of_symbol(const char *searched_sym);
-const char *find_sym_at_addr(ulong vaddr, ptrdiff_t *off, u32 *sym_size);
-const char *find_sym_at_addr_safe(ulong vaddr, ptrdiff_t *off, u32 *sym_size);
+const char *find_sym_at_addr(ulong vaddr, long *off, u32 *sym_size);
+const char *find_sym_at_addr_safe(ulong vaddr, long *off, u32 *sym_size);
 
 int foreach_symbol(int (*cb)(struct elf_symbol_info *, void *), void *arg);
